feat: Update model precision settings to bfloat16 and improve parameter filtering in adapters

+++ backend/config.py
@@ -17,7 +17,7 @@ THUMBNAIL_CACHE_MAX_SIZE = 100
 
 PRECISION_DEFAULTS = {
     'r4b': {'precision': 'float32', 'use_flash_attention': False},
-    'blip2': {'precision': 'float16', 'use_flash_attention': False},
+    'blip2': {'precision': 'bfloat16', 'use_flash_attention': False},
     'lfm2-vl-3b': {'precision': 'bfloat16', 'use_flash_attention': False},
     'llava-phi3': {'precision': 'float16', 'use_flash_attention': False},
     'nanonets-ocr-s': {'precision': 'bfloat16', 'use_flash_attention': False}
